requirements.txt
wheel
pillow
autopep8
pre_commit
fake_bpy_module_latest
qiskit
qiskit_finance
qiskit_machine_learning